D^V

+++ main-test/calico-test/multi-tenancy-generator/updatePools.py
@@ -17,7 +17,7 @@ spec:
 '''
 newPool=""
 for i in range(1, numOfTenants+1):
-    second=16+i/256
+    second=16+int(i/256)
     third=i%256
     newPool = template.format(id=i, secondDigit=second, thirdDigit=third)
     f = open("../pool/pools-" + str(i), "w")
